Adds Cruds views test.
