+++ app-engine/nose.cfg
@@ -1,7 +1,7 @@
 [nosetests]
 # Explicity include test directories here so I can avoid __init__.py files in test modules
 # and conflicts with app models in test_helper module.
-include=controllers|forms|models|handlers
+include=controllers|forms|models|handlers|views
 
 exclude-dir=lib
 verbosity=2
