+++ app-engine/templates/cruds/index.html
@@ -1,7 +1,7 @@
 {% extends 'layouts/base.html' %}
 
 {% block content %}
-<div id="cruds-index">
+<div id="cruds-index" class="cruds">
   <div class="header">
     <h1>Recent Cruds</h1>
