+++ app-engine/templates/cruds/new.html
@@ -4,7 +4,7 @@
 {% block title %}New Crud{% endblock %}
 
 {% block content %}
-<div id="crud-new" class="crud">
+<div id="cruds-new" class="cruds">
     <h2>Create a Crud</h2>
 
     <div class="row">
