+++ app-engine/templates/cruds/show.html
@@ -8,7 +8,7 @@
 
 
 {% block content %}
-    <div id="crud-show" class="cruds row">
+    <div id="cruds-show" class="cruds row">
         <div class="main-column col-sm-9">
             <h2>Crud</h2>
