+++ app-engine/tests/controllers/test_cruds_controller.py
@@ -37,63 +37,6 @@ class CrudsControllerTest(AppEngineControllerTest):
         self.assertEqual(response.status_code, 200, html)
         self.assertEqual(len(page), 1, html)
 
-    def test_expects_index_to_not_show_create_button_to_unauthenticated_user(self):
-        # Arrange
-        client = cruds_controller.test_client()
-        guest = MockIdentityService.unauthenticated_guest(self)
-        endpoint = '/cruds/'
-        button_selector = 'a.btn.create-crud'
-
-        # Assume
-        self.assertFalse(guest.is_authenticated())
-
-        # Act
-        response = client.get(endpoint, follow_redirects=False)
-        html = parse_html(response.data)
-        button = html.select(button_selector) if html else None
-
-        # Assert
-        self.assertEqual(response.status_code, 200, html)
-        self.assertEqual(len(button), 0, html)
-
-    def test_expects_index_to_not_show_create_button_to_authenticated_user(self):
-        # Arrange
-        client = cruds_controller.test_client()
-        guest = MockIdentityService.login_app_engine_user(self)
-        endpoint = '/cruds/'
-        button_selector = 'a.btn.create-crud'
-
-        # Assume
-        self.assertTrue(guest.is_authenticated())
-
-        # Act
-        response = client.get(endpoint, follow_redirects=False)
-        html = parse_html(response.data)
-        button = html.select(button_selector) if html else None
-
-        # Assert
-        self.assertEqual(response.status_code, 200, html)
-        self.assertEqual(len(button), 0, html)
-
-    def test_expects_index_to_show_create_button_to_admin_user(self):
-        # Arrange
-        client = cruds_controller.test_client()
-        admin = MockIdentityService.login_app_engine_user(self, as_admin=True)
-        endpoint = '/cruds/'
-        button_selector = 'a.btn.create-crud'
-
-        # Assume
-        self.assertTrue(admin.is_admin())
-
-        # Act
-        response = client.get(endpoint, follow_redirects=False)
-        html = parse_html(response.data)
-        button = html.select(button_selector) if html else None
-
-        # Assert
-        self.assertEqual(response.status_code, 200, html)
-        self.assertEqual(len(button), 1, html)
-
     def test_expects_admin_to_create_crud(self):
         # Arrange
         client = cruds_controller.test_client()
@@ -140,8 +83,36 @@ class CrudsControllerTest(AppEngineControllerTest):
         self.assertEqual(response.status_code, 403, response.data)
         self.assertEqual(html.h2.text, 'Permission Required')
 
-    def test_expects_unauthenticated_user_to_be_denied_new_view(self):
-        pass
+    def test_expects_admin_user_to_access_new_view(self):
+        # Arrange
+        client = cruds_controller.test_client()
+        admin = MockIdentityService.login_app_engine_user(self, as_admin=True)
+        endpoint = '/cruds/new/'
+        page_selector = 'div#cruds-new'
 
-    def test_expects_authenticated_user_to_access_new_view(self):
-        pass
+        # Assume
+        self.assertTrue(admin.is_admin())
+
+        # Act
+        response = client.get(endpoint, follow_redirects=False)
+        html = parse_html(response.data)
+        page = html.select(page_selector) if html else None
+
+        # Assert
+        self.assertEqual(response.status_code, 200, response.data)
+        self.assertEqual(len(page), 1, html)
+
+    def test_expects_unauthenticated_user_to_be_denied_new_view(self):
+        # Arrange
+        client = cruds_controller.test_client()
+        guest = MockIdentityService.unauthenticated_guest(self)
+        endpoint = '/cruds/new/'
+
+        # Assume
+        self.assertFalse(guest.is_authenticated())
+
+        # Act
+        response = client.get(endpoint, follow_redirects=False)
+
+        # Assert
+        self.assertEqual(response.status_code, 403, response.data)
